move the component to pd
