+++ components/src/organisms/index.ts
@@ -1,5 +1,4 @@
 export * from './DeckLabelSet'
-export * from './EndUserAgreementFooter'
 export * from './Toolbox'
 export * from './ProtocolTimelineScrubber'
 export * from './CommandText'
