+++ protocol-designer/src/assets/localization/en/shared.json
@@ -14,6 +14,7 @@
   "confirm": "Confirm",
   "consent_to_eula": "By using Protocol Designer, you consent to the Opentrons <link1>EULA</link1>.",
   "continue_with_export": "Continue with export",
+  "copyright": "Copyright © {{year}} Opentrons",
   "create_a_protocol": "Create a protocol",
   "create_new": "Create new",
   "destination_well": "Destination Well",
@@ -102,16 +103,17 @@
   "not_json": "Invalid file type",
   "one_channel": "1-Channel",
   "only_tiprack": "Incompatible file type",
+  "opentrons_collects_data": "In order to improve our products, Opentrons would like to collect data related to your use of Protocol Designer. Opentrons will collect and store analytics and session data, including through the use of cookies and similar technologies, solely for the purpose enhancing our products.",
   "opentrons_flex": "Opentrons Flex",
   "opentrons": "Opentrons",
-  "opentrons_collects_data": "In order to improve our products, Opentrons would like to collect data related to your use of Protocol Designer. Opentrons will collect and store analytics and session data, including through the use of cookies and similar technologies, solely for the purpose enhancing our products.",
   "ot2": "Opentrons OT-2",
   "overwrite_labware": "Overwrite labware",
   "overwrite": "Click Overwrite to replace the existing labware with the new labware.",
   "part": "Part {{current}} / {{max}}",
-  "pipette": "Pipette",
   "pd_version": "Protocol designer version",
+  "pipette": "Pipette",
   "primary_order": "Primary order",
+  "privacy_policy": "By continuing, you agree to the Opentrons <privacyPolicyLink>Privacy Policy</privacyPolicyLink> and <EULALink>End user license agreement</EULALink>",
   "privacy": "Privacy",
   "protocol_designer": "Protocol Designer",
   "re_export": "To use this definition, use Labware Creator to give it a unique load name and display name.",
@@ -119,19 +121,19 @@
   "release_notes": "Release notes",
   "reload_app": "Reload app",
   "remove": "remove",
-  "show_hints_and_tips": "Show all hints and tips notifications again",
-  "reset": "Reset",
   "reset_to_default": "Reset to default",
+  "reset": "Reset",
   "resize_your_browser": "Resize your browser to at least 768px wide and 650px tall to continue editing your protocol",
   "review_our_privacy_policy": "You can adjust this setting at any time by clicking on the settings icon. Find detailed information in our <link1>privacy policy</link1>.",
   "right": "Right",
   "save": "Save",
   "secondary_order": "Secondary order",
   "settings": "Settings",
+  "shared_analytics": "Share analytics with Opentrons",
   "shared_display_name": "Shared display name: ",
   "shared_load_name": "Shared load name: ",
-  "shared_analytics": "Share analytics with Opentrons",
   "shares_name": "This labware has the same load name or display name as {{customOrStandard}}, which is already in this protocol.",
+  "show_hints_and_tips": "Show all hints and tips notifications again",
   "slot_detail": "Slot Detail",
   "software_manual": "Software manual",
   "source_well": "Source Well",
