+++ protocol-designer/src/molecules/EndUserAgreementFooter/utils.ts
