+++ protocol-designer/src/pages/ProtocolOverview/index.tsx
@@ -9,7 +9,6 @@ import {
   ALIGN_CENTER,
   COLORS,
   DIRECTION_COLUMN,
-  EndUserAgreementFooter,
   Flex,
   JUSTIFY_FLEX_END,
   JUSTIFY_SPACE_BETWEEN,
@@ -31,6 +30,7 @@ import { actions as loadFileActions } from '../../load-file'
 import { selectors as labwareIngredSelectors } from '../../labware-ingred/selectors'
 import { MaterialsListModal } from '../../organisms/MaterialsListModal'
 import { LINE_CLAMP_TEXT_STYLE, COLUMN_STYLE } from '../../atoms'
+import { EndUserAgreementFooter } from '../../molecules'
 import { useBlockingHint } from '../../organisms/BlockingHintModal/useBlockingHint'
 import {
   EditProtocolMetadataModal,
